# docs/service/catalog-service/api.yaml
openapi: 3.0.3
info:
  title: catalog-service API
  version: "0.1"
  description: |
    Quản lý Catalog — Brand, Category, AttributeTemplate, Product (SPU), Variant (SKU).
    Upstream thuần túy: không consume event, chỉ publish.

    **Auth:**
    - `/api/**` — public (không cần token)
    - `/api/admin/**` — yêu cầu role `ADMIN` (bearer token từ oauth2-service)
servers:
  - url: http://localhost:8082
    description: Local dev
...
components:
  schemas:
    ApiResponseVoid:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          nullable: true
          example: null

    # ── Brand ─────────────────────────────────────────────────────────────────
...
    AttributeConstraints:
      type: object
      nullable: true
      description: >
        Ràng buộc theo ngành hàng, chồng lên guard mặc định của template (analysis.md §5.3.1). Field nào dùng
        phụ thuộc `inputType` — field không áp dụng phải để null, server 422 INVALID_CONSTRAINTS nếu sai:
        TEXT{maxLength} | NUMBER{min,max,integerOnly} | MULTI_SELECT{maxSelections}
        | DATE{inputPrecision,minDate,maxDate} | SINGLE_SELECT,BOOLEAN{ — không field nào}.
      properties:
        maxLength: { type: integer, nullable: true, description: "TEXT — ≤ 255" }
        min: { type: number, nullable: true, description: "NUMBER" }
        max: { type: number, nullable: true, description: "NUMBER" }
        integerOnly: { type: boolean, nullable: true, description: "NUMBER" }
        inputPrecision:
          nullable: true
          allOf: [{ $ref: '#/components/schemas/DatePrecision' }]
          description: "DATE — mức seller phải nhập tới"
        minDate: { type: string, nullable: true, description: "DATE — tuyệt đối (yyyy[-MM[-dd]]) hoặc tương đối (now-30d)" }
        maxDate: { type: string, nullable: true, description: "DATE" }
        maxSelections: { type: integer, nullable: true, description: "MULTI_SELECT — ≥ 1" }
    DatePrecision:
      type: string
      enum: [YEAR, MONTH, DAY]
    Discovery:
      type: object
      nullable: true
      description: >
        Cách 1 thuộc tính tham gia tìm hàng (analysis.md §5.3.2) — mỗi nhánh độc lập, null = tắt khả năng đó.
        Nhánh hợp lệ theo `inputType`: search (mọi kiểu trừ DATE), filter (không TEXT), sort (chỉ NUMBER/DATE).
      properties:
        search:
          type: string
          nullable: true
          enum: [NORMAL, HIGH]
          description: Trọng số full-text — không áp dụng cho DATE
        filter:
          $ref: '#/components/schemas/FilterConfig'
        sort:
          type: object
          nullable: true
          description: Chỉ NUMBER/DATE. Sản phẩm thiếu giá trị xếp cuối.
          properties:
            label: { type: string, description: "Nhãn hiển thị, VD \"Giá: Thấp đến cao\"" }
            direction: { type: string, enum: [ASC, DESC] }
    FilterConfig:
      type: object
      nullable: true
      required: [kind]
      description: >
        `kind` quyết định field nào có nghĩa: SELECT (SINGLE_SELECT/MULTI_SELECT) dùng match/valueOrder;
        BOOLEAN không field nào khác; RANGE (NUMBER/DATE) dùng presentation/step/buckets.
      properties:
        kind:
          type: string
          enum: [SELECT, BOOLEAN, RANGE]
        match:
          type: string
          nullable: true
          enum: [ANY, ALL]
          description: SELECT — SINGLE_SELECT chỉ nhận ANY, MULTI_SELECT nhận ANY/ALL
        valueOrder:
          type: string
          nullable: true
          enum: [OPTION_ORDER, COUNT]
          description: SELECT
        presentation:
          type: string
          nullable: true
          enum: [SLIDER, BUCKETS]
          description: RANGE
        step:
          nullable: true
          allOf: [{ $ref: '#/components/schemas/DatePrecision' }]
          description: RANGE — chỉ DATE + SLIDER; phải thô hơn hoặc bằng constraints.inputPrecision
        buckets:
          type: array
          nullable: true
          description: RANGE + BUCKETS — được chồng nhau, mỗi bucket cần ít nhất from hoặc to
          items:
            type: object
            required: [label]
            properties:
              label: { type: string }
              from: { type: string, nullable: true, description: "Mốc số (NUMBER) hoặc ngày tuyệt đối/tương đối VD now-30d (DATE)" }
              to: { type: string, nullable: true }
    CategoryAttributeResponse:
      type: object
      required: [templateId, name, displayName, inputType, required, displayOrder, options]
      description: >
        Thuộc tính seller dùng khi tạo/sửa sản phẩm ở ĐÚNG danh mục lá này (L1/L2 trả mảng rỗng). Chỉ template DÙNG
        ĐƯỢC (ACTIVE, và SELECT còn ≥ 1 option ACTIVE); `options` chỉ gồm option ACTIVE theo thứ tự chuẩn. KHÔNG
        kèm `discovery` (thuần cấu hình tìm kiếm, search-service đọc qua snapshot, không qua endpoint này).
      properties:
        templateId:
          type: string
          description: ULID
        name:
          type: string
          example: "color"
        displayName:
          type: string
          example: "Màu sắc"
        hint:
          type: string
          nullable: true
        inputType:
          $ref: '#/components/schemas/InputType'
        unit:
          type: string
          nullable: true
          example: "GB"
        required:
          type: boolean
        constraints:
          $ref: '#/components/schemas/AttributeConstraints'
        displayOrder:
          type: integer
        options:
          type: array
          description: Rỗng với kiểu không phải SELECT
          items:
            type: object
            required: [id, value, displayValue]
            properties:
              id: { type: string }
              value: { type: string }
              displayValue: { type: string }
    CategoryAssignmentDetailResponse:
      type: object
      required: [templateId, name, displayName, inputType, templateStatus, required, displayOrder]
      description: >
        Chi tiết quản trị 1 assignment — ĐỦ, kể cả template đã tắt (khác CategoryAttributeResponse chỉ trả
        template dùng được). Dùng để sửa (PUT lại toàn bộ) mà không làm rơi mất assignment cũ.
      properties:
        templateId: { type: string }
        name: { type: string }
        displayName: { type: string }
        inputType:
          $ref: '#/components/schemas/InputType'
        templateStatus:
          $ref: '#/components/schemas/AttributeTemplateStatus'
        required: { type: boolean }
        displayOrder: { type: integer }
        constraints:
          $ref: '#/components/schemas/AttributeConstraints'
        discovery:
          $ref: '#/components/schemas/Discovery'
    AssignableAttributeTemplateResponse:
      type: object
      required: [options]
      description: Droplist khi Admin gán thêm attribute cho danh mục lá — chỉ template dùng được, chưa gán.
      properties:
        options:
          type: array
          items:
            type: object
            required: [id, displayName, inputType]
            properties:
              id: { type: string }
              displayName: { type: string }
              inputType:
                $ref: '#/components/schemas/InputType'
    CreateCategoryRequest:
      type: object
      required: [name]
      description: "Không có slug — hệ thống tự sinh từ tên (bỏ dấu, chữ thường, gạch nối), không duy nhất, đổi theo tên. Tra cứu theo id (service.md T-10)."
      properties:
        name:
          type: string
          example: "Smartphones"
        parentId:
          type: string
          nullable: true
          description: "null = L1 root. Nếu parentId là L3 → 422 MAX_DEPTH_EXCEEDED"
    UpdateCategoryRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          example: "Smartphones & Tablets"
        imageUrl:
          type: string
          nullable: true
    ReorderCategoriesRequest:
      type: object
      required: [categoryIds]
      description: "`parentId` null = sắp lại category gốc (L1). Phải chứa MỌI anh em cùng cha, mỗi cái đúng 1 lần."
      properties:
        parentId:
          type: string
          nullable: true
        categoryIds:
          type: array
          items: { type: string }
          minItems: 1
    ReplaceCategoryAttributesRequest:
      type: object
      required: [templateId, required, displayOrder]
      description: >
        1 phần tử của body `PUT /api/admin/categories/{id}/attributes` — body là 1 ARRAY của schema này,
        thay THẾ TOÀN BỘ danh sách attribute hiện có của category (không phải merge/patch). Client tự
        GET `GetCategoryAttributeAssignments` (admin, trả đủ kể cả template đã tắt) trước, sửa trong list,
        rồi PUT lại nguyên list mong muốn cuối cùng — assign/update/remove 1 attribute đều quy về gọi lại
        endpoint này với list đã sửa. "Thêm mới" (templateId chưa có trong assignment hiện tại) phải là
        template dùng được; giữ/sửa assignment của template đã tắt vẫn được phép.
      properties:
        templateId:
          type: string
          description: ULID của AttributeTemplate
        required:
          type: boolean
        displayOrder:
          type: integer
          minimum: 0
        constraints:
          $ref: '#/components/schemas/AttributeConstraints'
        discovery:
          $ref: '#/components/schemas/Discovery'

    # ── AttributeTemplate ─────────────────────────────────────────────────────
    InputType:
      type: string
      enum: [SINGLE_SELECT, MULTI_SELECT, TEXT, NUMBER, BOOLEAN, DATE]
      description: >
        Bất biến. SINGLE/MULTI_SELECT có option (giá trị lưu là optionId). DATE lưu chuỗi ISO rút gọn
        (yyyy | yyyy-MM | yyyy-MM-dd). Quy tắc từng kiểu: analysis.md §5.3.1.
...
    AttributeTemplateStatus:
      type: string
      enum: [ACTIVE, INACTIVE]
...
  responses:
    400:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error:
                type: object
                properties:
                  code: { type: string, example: "VALIDATION_ERROR" }
                  message: { type: string }
    404:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error:
                type: object
                properties:
                  code: { type: string }
                  message: { type: string }
    409:
      description: Conflict (duplicate, in-use)
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error:
                type: object
                properties:
                  code: { type: string }
                  message: { type: string }
    422:
      description: Business rule violation
      content:
        application/json:
          schema:
            type: object
            properties:
              success: { type: boolean, example: false }
              error:
                type: object
                properties:
                  code: { type: string, example: "INPUT_TYPE_NOT_SELECT" }
                  message: { type: string }

    # ── Product ───────────────────────────────────────────────────────────────
...
paths:
...
  /api/categories/{id}/attributes:
    get:
      tags: [Category (Public)]
      summary: Lấy danh sách attributes của category
      description: |
        Chỉ template DÙNG ĐƯỢC được gán vào ĐÚNG danh mục lá này (không kế thừa giữa các level — L1/L2 trả
        mảng rỗng), sorted by displayOrder. Kết quả cache L2 (TTL 1hr) per categoryId.
      operationId: getCategoryAttributes
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: Category ULID
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CategoryAttributeResponse'
        "404":
          $ref: '#/components/responses/404'
          # ErrorCode: CATEGORY_NOT_FOUND

  # ── Category (Admin) ────────────────────────────────────────────────────────
  /api/admin/categories:
    post:
      tags: [Category (Admin)]
      summary: Tạo category mới
      description: |
        `parentId=null` → tạo L1 root.
        `parentId` của L3 category → 422 MAX_DEPTH_EXCEEDED.
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCategoryRequest'
      responses:
        "201":
          description: Created — trả về ULID của category mới
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data: { type: string, description: ULID }
        "400":
          $ref: '#/components/responses/400'
        "404":
          $ref: '#/components/responses/404'
          # ErrorCode: CATEGORY_NOT_FOUND (parentId không tồn tại)
        "409":
          $ref: '#/components/responses/409'
          # ErrorCode: CATEGORY_NAME_EXISTS_IN_PARENT
        "422":
          $ref: '#/components/responses/422'
          # ErrorCode: MAX_DEPTH_EXCEEDED
...
  /api/admin/categories/reorder:
    put:
      tags: [Category (Admin)]
      summary: Sắp lại thứ tự anh em cùng cha
      description: Gửi đủ mọi anh em cùng cha theo thứ tự mong muốn (`parentId=null` = category gốc L1).
      operationId: reorderCategories
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReorderCategoriesRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "422":
          $ref: '#/components/responses/422'
          # ErrorCode: INVALID_CATEGORY_ORDER
  /api/admin/categories/{id}:
    put:
      tags: [Category (Admin)]
      summary: Cập nhật tên và ảnh category
      operationId: updateCategory
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCategoryRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "400":
          $ref: '#/components/responses/400'
        "404":
          $ref: '#/components/responses/404'
        "409":
          $ref: '#/components/responses/409'
          # ErrorCode: CATEGORY_NAME_EXISTS_IN_PARENT
    delete:
      tags: [Category (Admin)]
      summary: Xóa category (hard delete)
      description: |
        Guard: có sub-category → 409 HAS_CHILDREN.
        Guard: đang được Product reference → 409 HAS_PRODUCT_REFERENCE.
        Xóa cascade: closure rows + attribute assignments + entity.
      operationId: deleteCategory
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "404":
          $ref: '#/components/responses/404'
        "409":
          $ref: '#/components/responses/409'
          # ErrorCode: HAS_CHILDREN | HAS_PRODUCT_REFERENCE
  /api/admin/categories/{id}/deactivate:
    post:
      tags: [Category (Admin)]
      summary: Deactivate category (soft toggle điều hướng, không phải xoá dữ liệu)
      description: |
        Không guard theo "đang có Product/children hay không" — khác hard-delete. Ẩn category khỏi
        `GET /api/categories` (cây); chặn `CreateProduct` mới dưới category này (422 CATEGORY_INACTIVE).
        Product cũ tham chiếu category này không bị ảnh hưởng gì.
      operationId: deactivateCategory
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "404":
          $ref: '#/components/responses/404'
  /api/admin/categories/{id}/activate:
    post:
      tags: [Category (Admin)]
      summary: Reactivate category đã deactivate
      operationId: activateCategory
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "404":
          $ref: '#/components/responses/404'
  /api/admin/categories/{id}/assignments:
    get:
      tags: [Category (Admin)]
      summary: Chi tiết assignment (quản trị) — ĐỦ, kể cả template đã tắt
      description: >
        Khác `GET /api/categories/{id}/attributes` (chỉ trả template dùng được, dành cho seller/buyer) —
        endpoint này trả MỌI assignment hiện có kèm trạng thái template, để admin sửa (PUT lại toàn bộ)
        không làm rơi mất assignment của template đã tắt.
      operationId: getCategoryAttributeAssignments
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CategoryAssignmentDetailResponse'
        "404":
          $ref: '#/components/responses/404'
  /api/admin/categories/{id}/assignable-templates:
    get:
      tags: [Category (Admin)]
      summary: Droplist template có thể gán thêm vào danh mục lá
      description: Chỉ template DÙNG ĐƯỢC và CHƯA gán vào danh mục này.
      operationId: listAssignableAttributeTemplates
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - { name: keyword, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: true }
                  data:
                    $ref: '#/components/schemas/AssignableAttributeTemplateResponse'
        "404":
          $ref: '#/components/responses/404'
  /api/admin/categories/{id}/attributes:
    put:
      tags: [Category (Admin)]
      summary: Thay toàn bộ danh sách attribute assignment của category (leaf/L3)
      description: |
        Body là 1 ARRAY, REPLACE TOÀN BỘ assignment hiện có — không phải merge/patch. Assign 1 attribute
        mới, sửa config 1 attribute, hay gỡ 1 attribute đều gọi endpoint này với list mong muốn cuối cùng
        (client tự GET `GET /api/admin/categories/{id}/assignments` trước, sửa trong list, rồi PUT lại).
        Category đích phải là leaf (L3) — không có kế thừa từ category cha. "Thêm mới" phải là template
        dùng được; giữ/sửa assignment của template đã tắt vẫn cho phép.
      operationId: replaceCategoryAttributeAssignments
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
          description: Category ULID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/ReplaceCategoryAttributesRequest'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseVoid'
        "400":
          $ref: '#/components/responses/400'
        "404":
          $ref: '#/components/responses/404'
          # ErrorCode: CATEGORY_NOT_FOUND | TEMPLATE_NOT_FOUND
        "422":
          $ref: '#/components/responses/422'
          # ErrorCode: ATTRIBUTE_ASSIGNMENT_REQUIRES_LEAF | TEMPLATE_INACTIVE (chỉ khi thêm mới)
          # | INVALID_CONSTRAINTS | INVALID_DISCOVERY | INVALID_RANGE_FILTER
        "409":
          $ref: '#/components/responses/409'
          # ErrorCode: ASSIGNMENT_ALREADY_EXISTS (trùng templateId trong cùng request)
          # ErrorCode: CATEGORY_NOT_FOUND | TEMPLATE_NOT_FOUND
        "409":
          $ref: '#/components/responses/409'
          # ErrorCode: ASSIGNMENT_ALREADY_EXISTS — 1 templateId xuất hiện nhiều hơn 1 lần trong request
        "422":
          $ref: '#/components/responses/422'
          # ErrorCode: ATTRIBUTE_ASSIGNMENT_REQUIRES_LEAF — chỉ category leaf (L3) mới assign được
          # ErrorCode: TEMPLATE_INACTIVE — không assign được template đã deactivate

  # ── Brand (Public) ──────────────────────────────────────────────────────────
...